refactor(flat-pages): move big classroom to flat pages
