+++ packages/flat-pages/.lintstagedrc.js
@@ -1,0 +1,6 @@
+const baseLintConfig = require("../../.lintstagedrc.js");
+
+module.exports = {
+    ...baseLintConfig,
+    "*.{ts,tsx}": ["eslint --cache --fix", "prettier --write"],
+};
