+++ packages/flat-pages/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "flat-pages",
+  "version": "0.1.0",
+  "license": "MIT",
+  "dependencies": {
+    "@ant-design/icons": "^4.7.0",
+    "@netless/app-countdown": "^0.0.2",
+    "@netless/app-geogebra": "^0.0.4",
+    "@netless/app-iframe-bridge": "^0.0.2",
+    "@netless/app-monaco": "^0.1.14-beta.1",
+    "@netless/app-plyr": "^0.1.3",
+    "@netless/app-slide": "^0.2.2",
+    "@netless/combine-player": "^1.1.6",
+    "@netless/fastboard-react": "^0.3.5-canary.5",
+    "@netless/flat-rtc": "workspace:*",
+    "@netless/flat-rtc-agora-web": "workspace:*",
+    "@netless/flat-rtm": "workspace:*",
+    "@netless/flat-rtm-agora": "workspace:*",
+    "@netless/flat-server-api": "workspace:*",
+    "@netless/flat-services": "workspace:*",
+    "@netless/flat-stores": "workspace:*",
+    "@netless/player-controller": "^0.0.9",
+    "@netless/video-js-plugin": "^0.3.8",
+    "@netless/window-manager": "^0.4.31",
+    "antd": "^4.19.2",
+    "axios": "^0.26.1",
+    "classnames": "^2.3.1",
+    "date-fns": "^2.28.0",
+    "flat-components": "workspace:*",
+    "flat-i18n": "workspace:*",
+    "flat-types": "workspace:*",
+    "i18next": "^21.6.14",
+    "i18next-browser-languagedetector": "^6.1.3",
+    "lodash-es": "^4.17.21",
+    "mobx": "^6.4.2",
+    "mobx-react-lite": "^3.3.0",
+    "polly-js": "^1.8.3",
+    "react": "^17.0.2",
+    "react-device-detect": "^2.1.2",
+    "react-dom": "^17.0.2",
+    "react-i18next": "^11.15.7",
+    "react-router-dom": "^5.2.0",
+    "react-router-last-location": "^2.0.1",
+    "react-use": "^17.3.2",
+    "side-effect-manager": "^0.1.5",
+    "uuid": "^8.3.2",
+    "value-enhancer": "^1.0.3",
+    "video.js": "7.10.2",
+    "white-web-sdk": "npm:white-web-sdk-esm@2.16.24"
+  },
+  "devDependencies": {
+    "@babel/standalone": "7.17.8",
+    "@netless/mini-svg-data-uri": "0.0.1",
+    "@types/loadable__component": "^5.13.4",
+    "@types/mime": "^2.0.3",
+    "dotenv": "^16.0.0",
+    "dotenv-expand": "^8.0.2",
+    "eslint-loader": "^4.0.2",
+    "mime": "^3.0.0"
+  }
+}
