+++ packages/flat-pages/src/components/CloudStorageButton.less
