+++ packages/flat-pages/src/components/Whiteboard.less
