+++ packages/flat-pages/src/utils/generate-avatar.ts
