+++ packages/flat-pages/src/utils/inject-cursor.ts
